e1000e: add PCI device id to enable support for 82567V-4
Patchwork-id: 26492
O-Subject: [PATCH RHEL6 11/12] e1000e: add PCI device id to enable support for
	82567V-4
Bugzilla: 607264
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>

Adding support for hardware that was added since our last major
backport.  Confirmed with Intel this upstream patch was fine to take
alone:

	commit 10df0b9116e2039d5585a196753e5f36d7afcba2
	Author: Bruce Allan <bruce.w.allan@intel.com>
	Date:   Mon May 10 15:02:52 2010 +0000

	    e1000e: add PCI device id to enable support for 82567V-4

Not tested as we do not have any of this hardware in the inventory.

This will resolve RHBZ 607264.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>