sync fcoe with upstream
Patchwork-id: 26499
O-Subject: [PATCH RHEL6]: sync fcoe
Bugzilla: 603263
RH-Acked-by: Rob Evers <revers@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This is for BZ 603263.

This patch is made over the last fcoe update here:
http://post-office.corp.redhat.com/archives/rhkernel-list/2010-June/msg00460.html
Those patches are now in the SCSI maintainer's tree.

This next patch brings in several bug fixes that were made as
a result of RHEL 6 testing. There is one cleanup patch to
fnic that I included just to be in sync with upstream.

Almost all the code in this patch is in the SCSI maintainers tree.
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=e105ce8b953e4fcf659b01a517c6e65b60b6834f
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=4a266f026a9938e3eb895fafde788c3958b3a55e
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=155272f41f27ff5b8c43517eb5bf95cde7154e3b
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=0f6aaeb7ac2c6067000d864637e9a2273f0822f5
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=e105ce8b953e4fcf659b01a517c6e65b60b6834f
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=5c53fa052127ac80a84f7be07555153cd6b74681
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=ff3c6e46b003c09d1381696d49927731ccf5c8c4
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=811774d3e1355e75e9c5ab9f897d75d779526faa
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=0bd3b646d7b9e951c25f2f7b203a95207adfd1b8
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=bd48929d6eaa37312545f71a7f337b564f0d6c2d
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=9fe78b2571268d05ff40244ba8765af6aab1e51f
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=9523402796d1bb9290798df68dc1b317cb27d2b2
This patch
http://www.open-fc.org/openfc/gitweb/?p=openfc/fcoe-next.git;a=commit;h=166d47945ec177433e32e1b692bf4339c0557cc1
is not yet upstream. It just gone done being tested by Cisco and Intel,
and will be sent to the linux-scsi list soon.

Intel and Cisco and Broadcom have tested the patches. They are
doing most of our fcoe testing for RHEL6. I have done some basic
tests like unplug cables and reset switches. I do not have FIP
to test some of them. I trust Cisco and Broadcom's testing on it.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>