Add IBoE support
Patchwork-id: 26500
O-Subject: [Patch RHEL6] Implement core IBoE support
Bugzilla: 571959

This patch adds the IBoE support infrastructure to the core of the IB
stack without actually enabling any drivers to support it.  Otherwise,
it is identical to the previous IBoE patch I mailed in last week.  The
only difference is that now we splitting out the driver enablement into
a separate patch.  We are using bug #571959 to track this patch.  All
other comments in previous thread apply.  Please ack.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>