hpsa: do not allow hard reset of 640x-boards
Patchwork-id: 26531
O-Subject: [RHEL6 PATCH6/6] hpsa: do not allow hard reset of 640x-boards
Bugzilla: 598681
RH-Acked-by: Neil Horman <nhorman@redhat.com>

This is for bz#598681

hpsa: forbid hard reset of 640x boards
The 6402/6404 are two PCI devices -- two Smart Array controllers
-- that fit into one slot.  It is possible to reset them independently,
however, they share a battery backed cache module.  One of the pair
controls the cache and the 2nd one access the cache through the first
one.  If you reset the one controlling the cache, the other one will
not be a happy camper.  So we just forbid resetting this conjoined
mess.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>