Fix vmalloc slow down
Patchwork-id: 26581
O-Subject: [RHEL6] Fix vmalloc slow down (bz #583026) - try #2
Bugzilla: 583026
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

Hi,

The following is Nick's updated patch which fixes the problem spotted
by Bob. Otherwise it is pretty much identical to the previous patch.
Apologies for not spotting that there was an updated version available
earlier.

I've run it through some basic tests and it seems to work ok. If we were
to run it through the specsfs tests again it is likely to delay things
considerably. This patch has been in -mm for some time now, and is about
to be submitted upstream.

Steve.

----------------------------------------------------------------------------
From: Nick Piggin <npiggin@suse.de>

Provide a free area cache for the vmalloc virtual address allocator, based
on the approach taken in the user virtual memory allocator.

This reduces the number of rbtree operations and linear traversals over
the vmap extents to find a free area.  The lazy vmap flushing makes this
problem worse because because freed but not yet flushed vmaps tend to
build up in the address space between flushes.

Steven noticed a performance problem with GFS2. Results are as follows...

Signed-off-by: Nick Piggin <npiggin@suse.de>
Reviewed-by: Minchan Kim <minchan.kim@gmail.com>
Cc: Steven Whitehouse <swhiteho@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>