hpsa: don't pretend the reset works
Patchwork-id: 26588
O-Subject: [RHEL6 PATCH7/6] hpsa: don't pretend the reset works
Bugzilla: 598681

This is for bz#598681

I've discussed the stale commands from previous kernel again with
Stephen Cameron(HP) and it turns out that:
"I think current versions of hpsa will, if an unexpected command
completion is detected, print a message, and do nothing else.  However, it's
possible that one of these stale commands could complete, and have the same
index as a command that the new kexec'ed kernel has sent down, in which case it
would look like the command has completed before it actually has, and then the
actual completion, when it finally arrived, would appear to be a stale
completion.  Essentially, it would be possible for two commands to be out on
the controller at the same time with the same tag.

I think I've now been convinced that it's best to bring things to a halt as
soon as we determine that the controller reset has not worked."

This patch returns ENODEV for the controllers which we were not able to reset properly.
The reset should switch the controller to 'normal' but it remains in performant mode.
It has been acked by Stephen.

Tomas

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>