ext4: stop issuing discards if not supported by device
Patchwork-id: 26591
O-Subject: [PATCH RHEL6] ext4: stop issuing discards if not supported by device
Bugzilla: 608731
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>
RH-Acked-by: Jerome Marchand <jmarchan@redhat.com>

For BZ 608731, straight backport of the following upstream commit.

Thanks,
-Eric

From: Eric Sandeen <sandeen@redhat.com>
Date: Sun, 16 May 2010 07:00:00 +0000 (-0400)
Subject: ext4: stop issuing discards if not supported by device
X-Git-Tag: v2.6.35-rc1~39^2~29
X-Git-Url: http://git.kernel.org/?p=linux%2Fkernel%2Fgit%2Ftorvalds%2Flinux-2.6.git;a=commitdiff_plain;h=a30eec2a8650a77f7

ext4: stop issuing discards if not supported by device

Turn off issuance of discard requests if the device does
not support it - similar to the action we take for barriers.
This will save a little computation time if a non-discardable
device is mounted with -o discard, and also makes it obvious
that it's not doing what was asked at mount time ...

Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>