efi: Fill all reserved memmap entries if add_efi_memmap specified
Patchwork-id: 26592
O-Subject: Re: [Patch RHEL6.0] x86 efi: Fill all reserved memmap entries if
	add_efi_memmap specified.
Bugzilla: 607386
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>

BZ#607386 UV - Incomplete data in /proc/iomem & /sys/firmware/memmap
First Post

Patch is not upstream.

It is *REQUIRED* for kdump and kexec on really large systems
(5Tb of memory is sufficient to trigger the problem).
If /proc/iomem is not complete the kexec command is not going to
inform the crash kernel about all of memory, so any system dump will
not be complete.  Not all of memory is written to a dump file, but
the makedumpfile command will select pages from the memory that the
crash kernel presents as /proc/vmcore.  And that memory will be
incomplete.

Currently, the e820_reserve_resources() function does not add entries
obtained via the "add_efi_memmap" kernel cmdline option.  This causes
/sys/firmware/memmap/... to be incomplete (stops after 128 entries).
Utilities that examine these entries then do not get the complete
picture of system memory.

This patch causes the above function to use the e820 memmap instead
of the e820_saved memmap if "add_efi_memmap" cmdline option is
specified.

Signed-off-by: Mike Travis <travis@sgi.com>
Signed-off-by: Jack Steiner <steiner@sgi.com>

v2:
> >
> >  I didn't know bootloader had a problem two years ago  And as I mentioned,
> >  I'm open to suggestions but what I've heard is "fix the bootloader". What
> >  I don't know is how to do that.
> >
> >  I see that BIOS sets up an e820 compatible memory map and extra memmap
> >  entries "somewhere".  EFI seems to pick this up via ELILO and somehow
> >  the kernel finds these extra memmap entries.
> >
> >  So my question is what is broken?  Should the bootloader not use the
> >  standard e820 memmap at all, and invent another way of passing the
> >  memmap entries?  And what about EFI grub, shouldn't that also change?
> >  And how does this affect compatibility?
> >
> >  I'm in the dark here as to how the bootloader manages to pass these
> >  entries from BIOS to the kernel, and how far I can go modifying these
> >  things when I don't completely understand them.  The last problem I
> >  worked on in ELILO I couldn't even print messages as that invalidated
> >  the "magic key" and it would loop endlessly between ELILO and BIOS.
> >
> >  The other thing is priorities.  If something is working and not
> >  causing undue problems, it does not get higher priority than issues
> >  that cause complete system failure.  I've suggested a workable
> >  (maybe short term solution), that gets us past the boot, and on to
> >  the other potentially far more important problems.
> >
> >  And p.s. it's been way longer than two years...;-)   [the other thing
> >  to consider is we haven't had real hardware that long, and many, many,
> >  many problems have surfaced since then.  The memmap problem only
> >  showed up when we ran over 128 entries.  When has that happened before?
> >  And how was it handled then?]
> >
>
The 128-entry limit (caused by the fixed array in struct setup_info) was
brought to our attention in 2007.  At that time, a number of
alternatives were discussed; Linus in particular was adamant that this
should be handled in the bootloader, and not by the kernel.

The result was that we added a facility to the kernel to supply
arbitrary information via a linked list.  That has been in there since
at least early 2008.  However, noone who cares about this issue has
bothered stepping up and put the support into the bootloaders --
instead, we got suckered into accepting a "temporary hack" which only
has had the effect of giving you guys yet another excuse not to do what
you signed off to do in the first place.

Accordingly, the right answer for me to do is to veto this and any
further patches which enable you to skirt your responsibility, and I'm
getting damned close to doing so.