avoid THP expose VM bugs
Patchwork-id: 26615
O-Subject: [PATCH RHEL6] avoid THP expose VM bugs
Bugzilla: 606131
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

I'd like this to be applied to RHEL6 especially now that a new bug has
been opened https://bugzilla.redhat.com/show_bug.cgi?id=609071 .

I can't risk VM bugs that exists regardless of THP to be exposed by
THP and then we don't know if the regression is because of THP or
not. In theory the corruption only triggers when ksm_does_need_to_copy
runs and it also goes away with the patch I posted yesterday but
there's no good explanation for that. I even tried to use mmap_sem
write mode in the page fault (which should prevent all races that were
fixed by yesterday patch) and it still happens. So despite the
do_swap_page races are real and can trigger (by unlocking swapcache
before copy_page runs in ksm_does_need_to_copy) there's a bigger can
of worms in there. I also now don't exclude the race could be mmu
notifier related. In the meantime we've to stop this to trigger on
beta testing or THP takes the blame. The below is also the right way
to go for the long run as we can't ignore the precious referenced bits
and screw the VM working set whenever a THP allocation happens. So
please apply to RHEL6.

Do not worry that I keep debugging this, and to debug this I've to
return PAGEREF_RECLAIM unconditionally (so I've to patch it anyway)
and then I only debug it with THP off to reduce to the minimum the
amount of code possibly involved in the race. It provides me no
benefit if people can still reproduce on RHEL6 because I'm already
aware of the problem and I can reproduce it here (now in seconds and
not hours after returning PAGEREF_RECLAIM unconditionally and not only
when a THP allocation triggers).

With this patch KSM works fine even with ksm_does_need_to_copy and THP
on of course (just like it works fine when THP is off, as it's totally
unrelated to THP).

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>