SUNRPC: Fix a re-entrancy bug in xs_tcp_read_calldir()
Patchwork-id: 26623
O-Subject: [RHEL6 PATCH 4/7] SUNRPC: Fix a re-entrancy bug in
	xs_tcp_read_calldir()
Bugzilla: 560786
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Trond Myklebust <Trond.Myklebust@netapp.com>

If the attempt to read the calldir fails, then instead of storing the read
bytes, we currently discard them. This leads to a garbage final result when
upon re-entry to the same routine, we read the remaining bytes.

Fixes the regression in bugzilla number 16213. Please see
    https://bugzilla.kernel.org/show_bug.cgi?id=16213

Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Cc: stable@kernel.org

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>