nouveau: disable acceleration on NVA3/NVA5/NVA8 by default
Patchwork-id: 26633
O-Subject: [PATCH] drm/nouveau: disable acceleration on NVA3/NVA5/NVA8 by default
Bugzilla: 591062
RH-Acked-by: Adam Jackson <ajax@redhat.com>

rhbz#591062

There's an GPU lockup problem for which the cause is currently unknown
on these chipsets.

Until it's resolved, it's better to leave the user with a working system
without acceleration than to have random lockups.

With this patch, acceleration will be off by default if a known problem
chipset is detected, but can be re-enabled with nouveau.noaccel=0 on
the kernel commandline.

Signed-off-by: Ben Skeggs <bskeggs@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>