dm ioctl: remove __dev_status from geometry and target message
Patchwork-id: 26647
O-Subject: [RHEL6 PATCH 1/3] dm ioctl: remove __dev_status from geometry and
	target message
Bugzilla: 609591
RH-Acked-by: Alasdair G Kergon <agk@redhat.com>

From: Peter Rajnoha <prajnoha@redhat.com>

BZ 609591

Remove useless __dev_status call while processing an ioctl that sets up
device geometry and target message.  The data is not returned to
userspace so there is no point collecting it and in the case of
target_message it is collected before processing the message so if it
did return it might be stale.

Signed-off-by: Peter Rajnoha <prajnoha@redhat.com>
Signed-off-by: Alasdair G Kergon <agk@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>