zfcp: Do not wait for SBALs on stopped queue
Patchwork-id: 26676
O-Subject: [RHEL6 PATCH 1/5] [s390x] zfcp: Do not wait for SBALs on stopped queue
Bugzilla: 606359
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Description
-----------
Reading the FC host statistics on an offline FCP subchannel
results in a 5 seconds wait.  When issuing a non-I/O request,
zfcp waits for up to 5 seconds for a free SBAL on the request queue.
Only after getting a free SBAL or the 5 seconds expire, the waiting
process continues.

Also exit the wait when the request queue is not active.

Bugzilla
--------
BZ 606359
https://bugzilla.redhat.com/show_bug.cgi?id=606359

Upstream status of the patch
----------------------------
The patch will be upstream as of kernel version 2.6.35.
The patch has been sent upstream for inclusion into a -rc release:
http://kerneltrap.org/mailarchive/linux-scsi/2010/6/21/6885531

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>