zfcp: Do not escalate scsi eh after fast_io_fail_tmo fired
Patchwork-id: 26679
O-Subject: [RHEL6 PATCH 3/5] [s390x] zfcp: Do not escalate scsi eh after
	fast_io_fail_tmo fired
Bugzilla: 606365
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Description
-----------
When a remote port is set to blocked while the scsi eh is
running, the scsi eh is blocked until the dev_loss_tmo fires.
This can delay failover to another path in a multipathing setup.
In a multipath setup, the fast_io_fail_tmo is usually set to
a lower value for a fast failover.

The fc transport function fc_block_scsi_eh waits until
the dev_loss_tmo fire or the remote port is online again.

Change the fc_block_scsi_eh callback to exit when the
fast_io_fail_tmo fires and pass this information back to
scsi_eh. After the fast_io_fail_tmo fired, the pending
I/O has been terminated, and there is no need for the
scsi eh to run more recovery actions for commands on this
port. This is the zfcp patch for this problem, the patch
depends on the SCSI commit 2f2eb58762b4dcddfe25c90800323765c1257eca.

Bugzilla
--------
BZ 606365
https://bugzilla.redhat.com/show_bug.cgi?id=606365

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.35
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=a1dbfddd02d2b3e75f8750c68563e1d6eec16260

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>