zfcp: Remove SCSI device during unit_remove
Patchwork-id: 26678
O-Subject: [RHEL6 PATCH 4/5] [s390x] zfcp: Remove SCSI device during unit_remove
Bugzilla: 589278
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Description
-----------
When issuing the commands to delete a SCSI device and
then to remove the zfcp unit from a script, the zfcp unit
remove can fail.

The unit_remove will fail when the reference count of the
unit is not zero. When the SCSI device exists, it holds a
reference to the unit. The upstream commit
d9a9cdfb078d755e648d53ec25b7370f84ee5729 changed the
deletion of a SCSI to be run asynchronously from a workqueue.
With this change, the actual removal of the SCSI device
can run after the unit_remove in zfcp and the
unit_remove will fail.

Get a reference to the SCSI device from the unit_remove
function and remove the SCSI device from this function.
If the SCSI device has already been deleted earlier,
unit_remove cannot get the reference and does nothing.
If the removal of the SCSI device is running on two threads,
this is protected by the scan_mutex and the second one will
exit early.

Bugzilla
--------
BZ 589278
https://bugzilla.redhat.com/show_bug.cgi?id=589278

Upstream status of the patch
----------------------------
The patch will be sent upstream for kernel 2.6.36.
However, the upstream kernel versions >= 2.6.34 are more robust
because of improvements in the reference counting code in the
zfcp driver.  The problem occurs more often on older kernel
levels including 2.6.32.

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>