GFS2: fix BUG in gfs2_adjust_quota
Patchwork-id: 26721
O-Subject: [RHEL6.0 PATCH][GFS2] - Bug 603827 - GFS2: BUG in gfs2_adjust_quota
Bugzilla: 603827
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

HighMem pages on i686 do not get mapped to the buffer_heads and this was
causing a NULL pointer dereference when we were trying to memset page buffers
to zero.
We now use zero_user() that kmaps the page and directly manipulates page data.
This patch also fixes a boundary condition that was incorrect.

Resolves: rhbz#603827

Signed-off-by: Abhi Das <adas@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>