lpfc Update from 8.3.5.13 to 8.3.5.14 FC/FCoE
Patchwork-id: 26746
O-Subject: [RHEL6.0 PATCH] lpfc Update from 8.3.5.13 to 8.3.5.14 FC/FCoE
Bugzilla: 603808
RH-Acked-by: Mike Christie <mchristi@redhat.com>
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=603808

https://brewweb.devel.redhat.com/taskinfo?taskID=2570098

Testing:  Tested running dt for 2 days.  This was the first
attempt to run an extended 5 day data integrity test and
IO hung after the 2nd day of testing.  This issue is
captured in bz612202.  Several outstanding issues exist
and it is unclear that the IO hang witnessed while testing
this patch are related to it directly.  This is currently
under investigation.  Currently waiting for comment on
bz612202 from Emulex.

Also tested at Emulex.

This patch addresses bug fixes.

This patch includes the following changes:
* Changed version number to 8.3.5.14
* Fix LPFC_READ_REV mailbox command for BE
* Added driver Advanced Error Reporting (AER) support for SLI4 devices
* Fixed FCF discovery failure after swapping FCoE port (CR 102947)
* Fix RPI leak after a VFC shut/noshut storm (CR 102507)
* Allow enabling MSI-X interrupt with fewer vectors than requested on SLI4
  device
* Removed NOOP condition checking logic
* Prevent point to point discovery on a FCoE UCNA (CR 102844)
* Prevent unregistring of unused FCF when FLOGI is pending (CR 102642)
* Restrict scsi buffer allocation (CR 102643)
* Fixed accounting of allocated SCSI buffers when post sgl fail (CR 102644)

The corresponding upstream LPFC driver patchset was submitted upstream on
6/8/10:

[PATCH 0/4] lpfc 8.3.14: Update lpfc to revision 8.3.14

This patch set updates the lpfc driver to revision 8.3.14

The patches for 8.3.14 contain:
- PCI fixes and enhancements
- FCoE Discovery Fixes
- SCSI and SLI API fixes
- Update Driver version to 8.3.14

The bugfixes were cut against scsi-misc-2.6 w/ the 8.3.13 lpfc patches (posted
yesterday) applied

-- james s

  Signed-off-by: Alex Iannicelli <alex.iannicelli@emulex.com>
  Signed-off-by: James Smart <james.smart@emulex.com>

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>