dm: check that target supports discard just before submitting request
Patchwork-id: 26748
O-Subject: [RHEL6 PATCH 1/4] dm: check that target supports discard just before
	submitting request
Bugzilla: 612014
RH-Acked-by: Alasdair G Kergon <agk@redhat.com>
RH-Acked-by: Mikulas Patocka <mpatocka@redhat.com>

From: Mikulas Patocka <mpatocka@redhat.com>

BZ 612014

Because of a various race conditions, a discard request may be submitted
to a Device Mapper device even if the device advertises that it doesn't
support discards. (for example, the device is linear, dm advertises
discard support, discard request is constructed, the table is reloaded
with mirror, the discard is submitted to the mirror target that doesn't
support it).

To solve these problems we must check that the target supports discard
just before submitting a discard request to it.

Signed-off-by: Mikulas Patocka <mpatocka@redhat.com>
Acked-by: Mike Snitzer <snitzer@redhat.com>

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>