nv50: supply encoder disable() hook
Patchwork-id: 26790
O-Subject: [RHEL6 PATCH 3/6] drm/nv50: supply encoder disable() hook
Bugzilla: 598842
RH-Acked-by: Dave Airlie <airlied@redhat.com>

Allows us to remove a driver hack that used to be necessary to disable
encoders in certain situations before setting up a mode.  The DRM has
better knowledge of when this is needed than the driver does.

This fixes a number of display switching issues.

Signed-off-by: Ben Skeggs <bskeggs@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>