ath9k: cleanup init error path
Patchwork-id: 26822
O-Subject: [RHEL6 PATCH] ath9k: cleanup init error path
Bugzilla: 610224
RH-Acked-by: Stanislaw Gruszka <sgruszka@redhat.com>

BZ610224

FIPS testing created a situation where drivers would fail to register
their hardware with mac80211 (due to failure to init WEP -- see
other patch).  In the case of ath9k, it failed to recognize that
failure and allowed its probe request to report success.  This meant
the module would load successfully.  Then the attempt to unload that
module would result in a null pointer dereference.

The patch below is specific to RHEL6, but it is based on a portion of
an upstream patch that restructured the error handling in the device
init path for ath9k.  I verified this pacth in my local environment.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>