igb: drop support for UDP hashing w/ RSS
Patchwork-id: 26832
O-Subject: [RHEL 6.0 PATCH] igb: drop support for UDP hashing w/ RSS
Bugzilla: 613782
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

Bugzilla:
https://bugzilla.redhat.com/show_bug.cgi?id=613782

Description:
This change removes UDP from the supported protocols for RSS hashing. The
reason for removing this protocol is because IP fragmentation was causing a
network flow to be broken into two streams, one for fragmented, and one for
non-fragmented and this in turn was causing out-of-order issues.

Upstream Status:
http://git.kernel.org/?p=linux/kernel/git/davem/net-next-2.6.git;a=commitdiff;h=4478a9cdf007a0418755a8a4016af8352fb1c1f3

Brew Build:
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=2590942

Test Status:
Tested by customer.

  Stefan

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>