Allow for BCM5709S to dump vmcore via NFS
Patchwork-id: 26860
O-Subject: [RHEL6.0 PATCH] Allow for BCM5709S to dump vmcore via NFS
Bugzilla: 577809
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

bz577809
https://bugzilla.redhat.com/bugzilla/show_bug.cgi?id=577809
BCM5709S ethernet adapter cannot be used to dump vmcore by NFS

Description of problem:
When vmcore is dumped on a system with a 5709S, the
interface associated with the 5709S won't come up.

Solution:
Basically, the 5709S uses MSI-X in the normal kernel but
non-MSI-X in the kdump kernel. The switching from one mode
to the other without a system reset causes trouble (i.e.
MSI-X mode is still active).

Broadcom provided a pointer to the following upstream patch:
bnx2: Always enable MSI-X on 5709

This patch removes the somewhat arbitrary requirement that
MSI-X should only be used when there are more than one
CPU. This allows both the normal kernel and the kdump kernel
to use MSI-X so there is no switching and thus no trouble.

upstream commit:
3d5f3a7bbd06065b06c7f65f948437ded40255ec

Brew:
Successfully built in Brew (task 2576817)

Testing:
Red Hat QE tested this patch and did not find the problem to
manifest. Refer to the bz for details.

Acks would be appreciated. Thanks.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>