xfs: rename XFS_IGET_BULKSTAT to XFS_IGET_UNTRUSTED
Patchwork-id: 26886
O-Subject: [RHEL6.0 PATCH 3/4] xfs: rename XFS_IGET_BULKSTAT to
	XFS_IGET_UNTRUSTED
Bugzilla: 607031
RH-Acked-by: Christoph Hellwig <chellwig@redhat.com>
RH-Acked-by: Dave Chinner <dchinner@redhat.com>

BZ607031

Inode numbers may come from somewhere external to the filesystem
(e.g. file handles, bulkstat information) and so are inherently
untrusted. Rename the flag we use for these lookups to make it
obvious we are doing a lookup of an untrusted inode number and need
to verify it completely before trying to read it from disk.

Signed-off-by: Jiri Pirko <jpirko@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>