don't reassign to ROM res if it is not going to be enabled
Patchwork-id: 26894
O-Subject: [PATCH RHEL6 BZ612950] REPOST pci: don't reassign to ROM res if it is
	not going to be enabled
Bugzilla: 612950
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: John Feeney <jfeeney@redhat.com>

Repost of the patch to fix a superfluous space character at the beginning
of a line.

pci: don't reassign to ROM res if it is not going to be enabled

A ROM resource that doesn't fit should not cause us to try to re-assign
all the bus resources.  Nobody generally cares, and re-assigning is
going to just cause way more troubles than it tries to solve.

Signed-off-by: Yinghai Lu <yinghai@kernel.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>

RHBZ#:
======
https://bugzilla.redhat.com/show_bug.cgi?id=612950

A previous patch set to fix the allocation of MMIO resources was incomplete
and introduced a problem for the nVidia setup. This patch is required for
the nVidia KMS to initialize properly.

kABI Status:
============
No symbols were harmed.

Brew:
=====
Built on all platforms.
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=2590244

Upstream Status:
================
Commit 9a928660c9dcaff568c9d379655c5aa16fb981f8

Test Status:
============

Test by Marko Myllynen at Red Hat on HP 8530w.

"I'm happy to confirm that the kernel with the "pci: don't reassign to ROM res
if it is not going to be enabled" patch fixes this issue on HP 8530w!"

Peter Bogdanovic
pbogdano@redhat.com
IBM on-site partner

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>