radeon: check/restore sanity before doing anything else with GPU
Patchwork-id: 26900
O-Subject: [PATCH] [RHEL6.0 PATCH] radeon: check/restore sanity before doing
	anything else with GPU.
Bugzilla: 612767
RH-Acked-by: Jerome Glisse <jglisse@redhat.com>
RH-Acked-by: Adam Jackson <ajax@redhat.com>

bz 612767

upstream status: posted to dri-devel mailing list, will go into drm-next soon.

On systems using kexec, the new kernel is booted straight from the old kernel,
without any warning to the graphics driver. So the GPU is basically left as-is
in a running state, however the CPU side is completly reset.

Without stating the saneness of anyone using kexec on live systems, we should
at least try not to crash the GPU. This patch resets 3 registers to 0 that
could cause bad things to happen to the running system.

This allows kexec to work on a Power6/RN50 system.

Signed-off-by: Dave Airlie <airlied@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>