jbd2: Fix I/O hang in jbd2_journal_release_jbd_inode
Patchwork-id: 26909
O-Subject: [PATCH RHEL6.0 BZ607254] jbd2: Fix I/O hang in
	jbd2_journal_release_jbd_inode
Bugzilla: 607254
RH-Acked-by: Josef Bacik <josef@redhat.com>
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: David Howells <dhowells@redhat.com>

RHBZ#:
======
https://bugzilla.redhat.com/show_bug.cgi?id=607254

Description:
============
Brian King has been debugging a hang in jbd2_journal_release_jbd_inode
which is being seen on Power 6 systems quite a lot. When we get
in the hung state, all I/O to the disk in question gets blocked
where we stay indefinitely. Looking at the task list, I can see
we are stuck in jbd2_journal_release_jbd_inode waiting on a
wake up. I added some debug code to detect this scenario and
dump additional data if we were stuck in jbd2_journal_release_jbd_inode
for longer than 30 minutes. When it hit, I was able to see that
i_flags was 0, suggesting we missed the wake up.

This patch changes i_flags to be an unsigned long, uses bit operators
to access it, and adds barriers around the accesses. Prior to applying
this patch, we were regularly hitting this hang on numerous systems
in our test environment. After applying the patch, the hangs no longer
occur. Its still not clear to me why the j_list_lock doesn't protect us
in this path. It also appears a hang very similar to this was seen
in the past and then was no longer recreatable:

http://forum.soft32.com/linux/20090310-ext4-hangs-ftopict478916.html

Signed-off-by: Brian King <brking@linux.vnet.ibm.com>

RHEL Version Found:
===================
RHEL 6.0

kABI Status:
============
No symbols were harmed.

Brew:
=====
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=2600273

Upstream:
=========
it was posted and discussed here
http://www.spinics.net/lists/linux-ext4/msg19912.html

updated V2 to capture Josef's comments listed below:
This seems to be a bit overkill, you could probably just get away with

	clear_bit()
	smp_mb__after_clear_bit

V2 of the patch is here
http://www.spinics.net/lists/linux-ext4/msg19917.html

hasn't made git tree yet, but we believe it will make one soon.

Test Status:
============

Tested by the IBM ppc stress test team. We have had couple versions
of this patch and it has been holding up for the last 2 weeks.
Brian's V2 of this patch we have had under heavy ppc stress env.
for 12 hours on several ppc systems that have hit in the ~7 hour
timeframe. We will continue to test this patch on these systems for
72 hours which is the length of this stress test.

===============================================================
Steve Best
IBM on-site partner

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>