be2iscsi: Fix for 64K data
Patchwork-id: 26913
O-Subject: [PATCH RHEL 6] be2iscsi: Fix for 64K data
Bugzilla: 608795
RH-Acked-by: Rob Evers <revers@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This is for BZ 608795.

>From upstream:
   This patch fixes a wraparound issue when the length=65536 for an sge.
Since it was defined as a short, the length becomes zero

This patch was made and tested by Emulex who have been doing
the QE on the drive for us. I did some basic IO tests and it
fixed the issue for me too.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>