memcontrol: prevent endless loop with huge pages and near-limit group
Patchwork-id: 26924
O-Subject: [RHEL6 PATCH] memcg-THP-check-room-rhel6
Bugzilla: 608996
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

This should avoid an endless loop with memcg with THP if reclaim
fails. I asked Kame to review on the bugzilla.

https://bugzilla.redhat.com/show_bug.cgi?id=608996#c26
=====
Subject: memcontrol: prevent endless loop with huge pages and near-limit group

From: Johannes Weiner <hannes@cmpxchg.org>

If reclaim after a failed charging was unsuccessful, the limits are
checked again, just in case they settled by means of other tasks.

This is all fine as long as every charge is of size PAGE_SIZE, because
in that case, being below the limit means having at least PAGE_SIZE
bytes available.

But with transparent huge pages, we may end up in an endless loop
where charging and reclaim fail, but we keep going because the limits
are not yet exceeded, although not allowing for a huge page.

Fix this up by explicitely checking for enough room, not just whether
we are within limits.

Signed-off-by: Johannes Weiner <hannes@cmpxchg.org>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>