remove khugepaged young bit check
Patchwork-id: 26928
O-Subject: [RHEL6 PATCH] remove khugepaged young bit check
Bugzilla: 615381
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Avi Kivity <avi@redhat.com>
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

Here the alternative simpler and obviously safe approach to fix
https://bugzilla.redhat.com/show_bug.cgi?id=615381 . This simply
removes the heuristic. This alters the kabi as well as the previous
patch so we can fix it in 6.1 if we want.

Feel free to choose between the two patches. Also a third option is
not to fix anything and things will be ok even if khugepaged may not
be as effective as it could be when system goes low on memory and the
new VM code clears the young bits from pte and sptes (and page struct)
even before trying to unmap the pages.
==========
Subject: remove khugepaged young bit check

From: Andrea Arcangeli <aarcange@redhat.com>

But the spte young bits for guest access aren't available in pte_young
check that khugepaged does: the gup-fast effect of a EPT minor fault
also isn't available through the pte but it's available through the
page. Remove the heuristic to prevent khugepaged not to be as
effective as it could be. Register test_young placeholder in the mmu
notifier structure for the future.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>