Return EFAULT from kvm ioctl when guest access bad area
Patchwork-id: 26936
O-Subject: [PATCH RHEL6 1/2] Return EFAULT from kvm ioctl when guest access bad
	area.
Bugzilla: 608595
RH-Acked-by: Avi Kivity <avi@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>

Currently if guest access address that belongs to memory slot but is not
backed up by page or page is read only KVM treats it like MMIO access.
Remove that capability. It was never part of the interface and should
not be relied upon.

Signed-off-by: Gleb Natapov <gleb@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>